python_service/requirements.txt
dash==2.17.1
pandas==2.2.2
plotly==5.22.0
sqlalchemy==2.0.31
python-dotenv==1.0.1
dash-bootstrap-components==1.6.0
numpy==1.26.4
psycopg2-binary==2.9.9
scipy==1.13.1
pyspark==3.5.1
gunicorn==22.0.0
